requirements.txt
pandas
openai
sentence-transformers
faiss-cpu
Levenshtein
python-dotenv
fastapi
uvicorn
python-multipart
